refactor(backend): split public catalog distribution
