+++ apps/backend/src/catalog/distribution/public/index.ts
@@ -3,13 +3,17 @@ export {
   listPublicCatalogPackagesInExecutor,
   loadPublicCatalogPackageDetail,
   loadPublicCatalogPackageDetailInExecutor,
+  loadPublicCatalogPackageVersionCardPreview,
+  loadPublicCatalogPackageVersionCardPreviewInExecutor,
+  normalizeCatalogPublicPackageCardPreviewInput,
+  normalizeCatalogPublicPackageListInput,
+} from "./browse";
+export {
   loadPublicCatalogPackageMediaAssetsInExecutor,
   loadPublicCatalogPackageMediaForDownload,
   loadPublicCatalogPackageMediaForDownloadInExecutor,
-  loadPublicCatalogPackageVersionCardPreview,
-  loadPublicCatalogPackageVersionCardPreviewInExecutor,
+} from "./media";
+export {
   loadPublicCatalogSnapshot,
   loadPublicCatalogSnapshotInExecutor,
-  normalizeCatalogPublicPackageCardPreviewInput,
-  normalizeCatalogPublicPackageListInput,
-} from "./public";
+} from "./snapshot";
